UUDESTAAN taas logiikka.keko.* uudelleen nimetty duplicate class errorin takia
