+++ Tiralabra_maven/src/main/java/com/mycompany/Tiralabra_maven/logiikka/dijkstra/DijkstraWithHeap.java
@@ -2,7 +2,7 @@ package com.mycompany.Tiralabra_maven.logiikka.dijkstra;
 
 import com.mycompany.Tiralabra_maven.logiikka.Paikka;
 import com.mycompany.Tiralabra_maven.logiikka.Piste;
-import com.mycompany.Tiralabra_maven.logiikka.keko.JK;
+import com.mycompany.Tiralabra_maven.logiikka.keko.PriorityQueueKeko;
 import java.util.Stack;
 
 /**
@@ -43,7 +43,7 @@ public class DijkstraWithHeap {
 
         this.initialiseSingleSource();
 
-        JK heap = rakennaKekoJaAsetaVieruspaikat();
+        PriorityQueueKeko heap = rakennaKekoJaAsetaVieruspaikat();
 
         System.out.println(this.paikat[0][0].vierusPaikat);
 
@@ -86,8 +86,8 @@ public class DijkstraWithHeap {
         this.paikat[this.lahtoPiste.i][this.lahtoPiste.j].etaisyysAlkuun = 0;
     }
 
-    private JK rakennaKekoJaAsetaVieruspaikat() {
-        JK heap = new JK();
+    private PriorityQueueKeko rakennaKekoJaAsetaVieruspaikat() {
+        PriorityQueueKeko heap = new PriorityQueueKeko();
 
         for (int i = 0; i < this.paikat.length; i++) {
             for (int j = 0; j < this.paikat[0].length; j++) {
