+++ Tiralabra_maven/src/main/java/com/mycompany/Tiralabra_maven/logiikka/keko/MinKeko.java
@@ -1,69 +1,44 @@
 package com.mycompany.Tiralabra_maven.logiikka.keko;
 
 import com.mycompany.Tiralabra_maven.logiikka.Paikka;
-import java.util.PriorityQueue;
 
 /**
  *
  * @author Hannu
  */
-public class JK implements MK{
-    
-    private PriorityQueue<Paikka> q;
-    
-    public JK(){
-        this.q=new PriorityQueue<Paikka>();
-    }
-
+public interface MinKeko {
     /**
      * Kekoon lisätään kekoalkio.
      *
      * @param kekoAlkio lisättävä kekoalkio
      */
-    @Override
-    public void heapInsert(Paikka kekoAlkio) {
-        this.q.add(kekoAlkio);
-    }
+    public void heapInsert(Paikka kekoAlkio);
 
     /**
      * Keosta poistetaan huipulla oleva pienimmän avaimen sisältävä kekoalkio.
      *
      * @return keosta poistettu pienimmän avaimen sisältänyt kekoalkio
      */
-    @Override
-    public Paikka heapDelMin() {
-        return this.q.poll();
-    }
+    public Paikka heapDelMin();
 
     /**
      * Keosta poistetaan kekoalkio.
      *
      * @param kekoAlkio poistettava kekoalkio
      */
-    @Override
-    public void heapDelete(Paikka kekoAlkio) {
-        this.q.remove(kekoAlkio);
-    }
+    public void heapDelete(Paikka kekoAlkio);
 
     /**
      * Asettaa kekoalkion oikealle paikalle keossa, jos kekoalkion avain on muuttunut.
      *
      * @param kekoAlkio oikealle paikalle asetettava kekoalkio
      */
-    @Override
-    public void heapDecreaseKey(Paikka kekoAlkio) {
-        this.q.remove(kekoAlkio);
-        this.q.add(kekoAlkio);
-    }
+    public void heapDecreaseKey(Paikka kekoAlkio);
 
     /**
      * Tarkastaa onko keko tyhjä.
      *
      * @return palautetaan true, jos keko on tyhjä
      */
-    @Override
-    public boolean heapIsEmpty() {
-        return this.q.isEmpty();
-    }
-    
+    public boolean heapIsEmpty();
 }
