--- Tiralabra_maven/src/main/java/com/mycompany/Tiralabra_maven/logiikka/keko/NewClass.java
@@ -1,9 +1,0 @@
-package com.mycompany.Tiralabra_maven.logiikka.keko;
-
-/**
- *
- * @author Hannu
- */
-public class NewClass {
-    
-}
